Changing $LANDIS_CONSOLE to adapt to new location for v8
Thomas Estabrook and I had issues making v8 work because of strange libraries path reference issues. He managed to make it work, and one of his fixes was to move the LANDIS-II install from `/bin/LANDIS_Linux` to `/bin/LANDIS_Linux/Core-Model-v8-LINUX/`. Because of this, I'm changing the environment variable to point to the right place.