add tests to Clean_Docker_LANDIS-II_8_AllExtensions/Dockerfile
In order to run the tests as part of the image build process, we need to add a copy of the `tests/` folder to the `Clean_Docker_LANDIS-II_8_AllExtensions/` directory, since that folder is the docker build context for this image.
(These test files are small so it's probably not a big deal that they duplicate existing repo files.)

The tests are copied over and run after the landis build completes, and subsequently deleted (along with all test outputs) to keep them out of the final image.
However, this could be improved even further by using a multistage build in the future.